Replaced `$(window).on(...)` and `$(window).off(...)` with plain a JavaScript version.
What
=
Replaced `$(window).on(...)` and `$(window).off(...)` with plain JavaScript `addEventListener` and `removeEventListener` and added a polyfill for older browsers.

Why
=
This is being done to step away from jQuery and move to plain JavaScript. In plain JavaScript `window.addEventListener` is supported by modern browsers and IE9+ and the polyfill adds support for <IE9.