Added initialization as a global without jQuery.
### What
Added initialization as a global without jQuery. The library is always available on `window.scrollDepth` and can be initialized with `window.scrollDepth.init(options)`. If jQuery is available then the original `jQuery.scrollDepth(options)` is also available and functions the same.

### Why
The library no longer has any dependencies on jQuery and this allows it to be used independently of jQuery.

### Also
Moved the jQuery example to `index.jquery.html` and `index.html` is an example without jQuery.

This makes the library no longer dependent on jQuery and the dependency has been removed from `bower.json`.